apps/api/pyproject.toml
[tool.poetry]
name = "verigov-api"
version = "0.1.0"
description = "VeriGov API"
authors = ["VeriGov Dev <dev@verigov.local>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { version = "^0.27.1", extras = ["standard"] }
sqlalchemy = "^2.0.27"
alembic = "^1.13.1"
psycopg = { version = "^3.1.18", extras = ["binary"] }
pydantic = "^2.6.1"
pydantic-settings = "^2.2.1"
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
pyjwt = "^2.8.0"
httpx = "^0.27.0"
redis = "^5.0.1"
structlog = "^24.1.0"
python-multipart = "^0.0.9"
email-validator = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
